pyproject.toml
[project]
name = "llm-agent-book"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "rmc8", email = "k@rmc-8.com" }
]
dependencies = [
    "gradio>=5.17.1",
    "langchain-ollama>=0.2.3",
    "fire>=0.7.0",
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "langchain-openai>=0.3.7",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.6",
    "langchain-chroma>=0.2.2",
    "serpapi>=0.1.5",
    "google-search-results>=2.4.2",
    "langchainhub>=0.1.21",
    "mem0ai>=0.1.66",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/llm_agent_book"]
